+++ src/to_big_float.rs
@@ -3,7 +3,7 @@ use UpdateBigFloat;
 
 pub trait ToBigFloat {
     fn to_big_float(self) -> BigFloat;
-    fn to_big_float_with_prec(self, precision: uint) -> BigFloat;
+    fn to_big_float_with_prec(self, precision: usize) -> BigFloat;
 }
 
 impl<T: UpdateBigFloat> ToBigFloat for T {
@@ -13,7 +13,7 @@ impl<T: UpdateBigFloat> ToBigFloat for T {
     }
 
     #[inline]
-    fn to_big_float_with_prec(self, precision: uint) -> BigFloat {
+    fn to_big_float_with_prec(self, precision: usize) -> BigFloat {
         BigFloat::new().with_prec(precision).from(self)
     }
 }
